src/main/java/org/bitbucket/rwitzel/couchrepositoryexample/BlogPost.java
package org.bitbucket.rwitzel.couchrepositoryexample;

import java.util.List;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonInclude.Include;

/**
 * This entity represents a blog post.
 */
@JsonInclude(Include.NON_NULL)
public class BlogPost {

    private String type = "BlogPost";

    private String id;

    private String revision;

    private List<String> tags;

    @JsonProperty("_id")
    public String getId() {
        return id;
    }

    @JsonProperty("_id")
    public void setId(String id) {
        this.id = id;
    }

    @JsonProperty("_rev")
    public String getRevision() {
        return revision;
    }

    @JsonProperty("_rev")
    public void setRevision(String revision) {
        this.revision = revision;
    }

    public List<String> getTags() {
        return tags;
    }

    public void setTags(List<String> tags) {
        this.tags = tags;
    }

    public String getType() {
        return type;
    }

}
